--- java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/offerset/OfferSetRequestParserFactory.java
@@ -1,17 +1,0 @@
-/**
- * 
- */
-package net.ivoa.calycopis.offerset;
-
-import net.ivoa.calycopis.factory.FactoryBase;
-
-/**
- * 
- */
-public interface OfferSetRequestParserFactory
-    extends FactoryBase
-    {
-
-    public OfferSetRequestParser create();
-    
-    }
